Enable clipboard on GTK (port changes from stable branch)
Also deleted unused Clipboard.gtk.cs
